+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### ? - ?
+
+##### Fixes :wrench:
+
+- Fixed how the occlusion strength is used in the default tileset shader which was causing shadows to be too dark.
+
 ### v1.3.1 - 2023-06-06
 
 ##### Fixes :wrench: