+++ CHANGES.md
@@ -2,8 +2,13 @@
 
 ### ? - ?
 
+##### Additions :tada:
+
+- KTX2 compressed textures now remain compressed all the way to the GPU, reducing GPU memory usage.
+
 ##### Fixes :wrench:
 
+- Fixed how the occlusion strength is used in the default tileset shader which was causing shadows to be too dark.
 - Fixed a bug that caused a prefab with a `CesiumGlobeAnchor` to lose its position after save/reload.
 
 ### v1.3.1 - 2023-06-06